use ufunc to improve adjustment speed
