+++ cmethods/__init__.py
@@ -23,12 +23,10 @@ r"""
 
 from __future__ import annotations
 
-import multiprocessing
-from typing import Any, Callable, List, Optional, Union
+from typing import Any, List, Optional, Union
 
 import numpy as np
 import xarray as xr
-from tqdm import tqdm
 
 __author__ = "Benjamin Thomas Schwertfeger"
 __copyright__ = __author__
@@ -36,357 +34,28 @@ __email__ = "contact@b-schwertfeger.de"
 __link__ = "https://github.com/btschwertfeger"
 __github__ = "https://github.com/btschwertfeger/python-cmethods"
 
+from cmethods.static import ADDITIVE, METHODS, MULTIPLICATIVE
+from cmethods.utils import (
+    UnknownMethodError,
+    check_types,
+    ensure_devidable,
+    get_adjusted_scaling_factor,
+    get_cdf,
+    get_inverse_of_cdf,
+    nan_or_equal,
+)
 
-class UnknownMethodError(Exception):
-    """
-    Exception raised for errors if unknown method called in CMethods class.
-    """
 
-    def __init__(self: UnknownMethodError, method: str, available_methods: list):
-        super().__init__(
-            f'Unknown method "{method}"! Available methods: {available_methods}'
-        )
-
-
-class CMethods:
-    """
-     The CMethods class serves a collection of bias correction procedures to adjust
-     time-series of climate data.
-
-     The following bias correction techniques are available:
-         Scaling-based techniques:
-             * Linear Scaling :func:`cmethods.CMethods.linear_scaling`
-             * Variance Scaling :func:`cmethods.CMethods.variance_scaling`
-             * Delta (change) Method :func:`cmethods.CMethods.delta_method`
-
-         Distribution-based techniques:
-             * Quantile Mapping :func:`cmethods.CMethods.quantile_mapping`
-             * Detrended Quantile Mapping :func:`cmethods.CMethods.detrended_quantile_mapping`
-             * Quantile Delta Mapping :func:`cmethods.CMethods.quantile_delta_mapping`
-
-     Except for the Variance Scaling all methods can be applied on both, stochastic and non-stochastic
-     variables. The Variance Scaling can only be applied on stochastic climate variables.
-
-    - Non-stochastic climate variables are those that can be predicted with relative certainty based
-     on factors such as location, elevation, and season. Examples of non-stochastic climate variables
-     include air temperature, air pressure, and solar radiation.
-
-     - Stochastic climate variables, on the other hand, are those that exhibit a high degree of
-       variability and unpredictability, making them difficult to forecast accurately.
-       Precipitation is an example of a stochastic climate variable because it can vary greatly in timing,
-       intensity, and location due to complex atmospheric and meteorological processes.
-    """
-
-    SCALING_METHODS: List[str] = ["linear_scaling", "variance_scaling", "delta_method"]
-    DISTRIBUTION_METHODS: List[str] = [
-        "quantile_mapping",
-        "detrended_quantile_mapping",
-        "quantile_delta_mapping",
-    ]
-
-    CUSTOM_METHODS: List[str] = SCALING_METHODS + DISTRIBUTION_METHODS
-    METHODS: List[str] = CUSTOM_METHODS
-
-    ADDITIVE: List[str] = ["+", "add"]
-    MULTIPLICATIVE: List[str] = ["*", "mult"]
-
+class CMethods2:
     MAX_SCALING_FACTOR: Union[int, float] = 10
-
-    @classmethod
-    def get_available_methods(cls: CMethods) -> List[str]:
-        """
-        Function to return the available adjustment methods of the CMethods class.
-
-        :return: List of available bias correction methods
-        :rtype: List[str]
-
-        .. code-block:: python
-            :linenos:
-            :caption: Example: Get available methods
-
-            >>> from cmethods import CMethods as cm
-
-            >>> cm.get_available_methods()
-            [
-                "linear_scaling", "variance_scaling", "delta_method",
-                "quantile_mapping", "quantile_delta_mapping"
-            ]
-        """
-        return cls.METHODS
-
-    @classmethod
-    def get_function(cls: CMethods, method: str) -> Callable:
-        """
-        Returns the bias correction function corresponding to the ``method`` name.
-
-        :param method: The method name to get the function for
-        :type method: str
-        :raises UnknownMethodError: If the function is not implemented
-        :return: The function of the corresponding method
-        :rtype: function
-        """
-        if method == "linear_scaling":
-            return cls.linear_scaling
-        if method == "variance_scaling":
-            return cls.variance_scaling
-        if method == "delta_method":
-            return cls.delta_method
-        if method == "quantile_mapping":
-            return cls.quantile_mapping
-        if method == "detrended_quantile_mapping":
-            return cls.detrended_quantile_mapping
-        if method == "empirical_quantile_mapping":
-            return cls.empirical_quantile_mapping
-        if method == "quantile_delta_mapping":
-            return cls.quantile_delta_mapping
-        raise UnknownMethodError(method, cls.METHODS)
-
-    @classmethod
-    def adjust_3d(
-        cls: CMethods,
-        method: str,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
-        n_quantiles: int = 100,
-        kind: str = "+",
-        group: Optional[str] = None,
-        n_jobs: int = 1,
-        **kwargs: Any,
-    ) -> xr.core.dataarray.DataArray:
-        """
-        Function to apply a bias correction method on 3-dimensional climate data.
-
-        *It is very important to pass ``group="time.month`` for scaling-based
-        techniques if the correction should be performed as described in the
-        referenced articles!* It is wanted to be the default.
-
-        :param method: The bias correction method to use
-        :type method: str
-        :param obs: The reference data set of the control period
-            (in most cases the observational data)
-        :type obs: xr.core.dataarray.DataArray
-        :param simh: The modeled data of the control period
-        :type simh: xr.core.dataarray.DataArray
-        :param simp: The modeled data of the scenario period (this is the data set
-            on which the bias correction takes action)
-        :type simp: xr.core.dataarray.DataArray
-        :param n_quantiles: Number of quantiles to respect. Only applies to
-            distribution-based bias correction techniques, defaults to ``100``
-        :type n_quantiles: int, optional
-        :param kind: The kind of adjustment - additive or multiplicative, defaults to ``"+"``
-        :type kind: str, optional
-        :param group: The grouping base, Only applies to scaling-based techniques, defaults to ``None``
-        :type group: str, optional
-        :param n_jobs: Number of parallels jobs to run the correction, defaults to ``1``
-        :type n_jobs: int, optional
-        :raises UnknownMethodError: If the correction method is not implemented
-        :return: The bias-corrected time series
-        :rtype: xr.core.dataarray.DataArray
-
-        .. code-block:: python
-            :linenos:
-            :caption: Example application - 3-dimensional bias correction
-
-            >>> import xarray as xr
-            >>> from cmethods import CMethods as cm
-
-            >>> # Note: The data sets must contain the dimensions "time", "lat", and "lon"
-            >>> #       for the respective variable.
-            >>> obsh = xr.open_dataset("path/to/reference_data-control_period.nc")
-            >>> simh = xr.open_dataset("path/to/modeled_data-control_period.nc")
-            >>> simp = xr.open_dataset("path/to/the_dataset_to_adjust-scenario_period.nc")
-            >>> variable = "tas" # temperatures
-
-            '''
-            In the following the Quantile Delta Mapping techniques is applied on
-            3-dimensional time-series data sets containing the variable "tas". The
-            parameter "kind" is not specified, since the additive ("+") kind is the default.
-            When adjusting ratio-based variables like precipitation, the multiplicative
-            variant ("*") should be used. In addition "n_jobs" is set to 4 which means that
-            four processes are used. This can improve the overall execution time.
-
-            After the execution, "qdm_adjusted" contains the fully bias-corrected
-            data - with the same shape, resolution, dimensions, coordinates and
-            attributes.
-            '''
-            >>> qdm_adjusted = cm.adjust_3d(
-            ...     method = "quantile_delta_mapping",
-            ...     obs = obs[variable],
-            ...     simh = simh[variable],
-            ...     simp = simp[variable],
-            ...     n_quantiles = 250,
-            ...     n_jobs = 4
-            ... )
-
-            '''
-            The next example shows how to apply the Linear Scaling bias correction
-            technique based on long-term monthly means.
-            '''
-            >>> ls_adjusted = cm.adjust_3d(
-            ...     method="linear_scaling",
-            ...     obs=obs[variable],
-            ...     simh=simh[variable],
-            ...     simp=simp[variable],
-            ...     group="time.month",
-            ...     kind="+"
-            ... )
-        """
-
-        if not isinstance(obs, xr.core.dataarray.DataArray):
-            raise TypeError("'obs' must be type xarray.core.dataarray.DataArray")
-        if not isinstance(simh, xr.core.dataarray.DataArray):
-            raise TypeError("'simh' must be type xarray.core.dataarray.DataArray")
-        if not isinstance(simp, xr.core.dataarray.DataArray):
-            raise TypeError("'simp' must be type xarray.core.dataarray.DataArray")
-        if not isinstance(n_quantiles, int):
-            raise TypeError("'n_quantiles' must be type int")
-        if not isinstance(n_jobs, int):
-            raise TypeError("'n_jobs' must be type int")
-
-        obs = obs.transpose("lat", "lon", "time")
-        simh = simh.transpose("lat", "lon", "time")
-        simp = simp.transpose("lat", "lon", "time").load()
-
-        result = simp.copy(deep=True)
-        len_lat, len_lon = len(simp.lat), len(simp.lon)
-
-        if method in cls.CUSTOM_METHODS:
-            if n_jobs == 1:
-                func = cls.get_function(method)
-                for lat in tqdm(range(len_lat)):
-                    for lon in range(len_lon):
-                        result[lat, lon] = func(
-                            obs=obs[lat, lon],
-                            simh=simh[lat, lon],
-                            simp=simp[lat, lon],
-                            group=group,
-                            kind=kind,
-                            n_quantiles=n_quantiles,
-                            **kwargs,
-                        )
-            else:
-                try:
-                    pool = multiprocessing.Pool(processes=n_jobs)
-                    # with multiprocessing.Pool(processes=n_jobs) as pool:
-                    # context manager is not used because than the coverage does not work.
-                    # this may change when upgrading to only support Python 3.11+
-                    params: List[dict] = [
-                        {
-                            "method": method,
-                            "obs": obs[lat],
-                            "simh": simh[lat],
-                            "simp": simp[lat],
-                            "group": group,
-                            "kind": kind,
-                            "n_quantiles": n_quantiles,
-                            "kwargs": kwargs,
-                        }
-                        for lat in range(len_lat)
-                    ]
-                    for lat, corrected in enumerate(pool.map(cls.pool_adjust, params)):
-                        result[lat] = corrected
-                finally:
-                    pool.close()
-                    pool.join()
-
-            return result.transpose("time", "lat", "lon")
-        raise UnknownMethodError(method, cls.METHODS)
-
-    @classmethod
-    def pool_adjust(cls: CMethods, params: dict) -> np.ndarray:
-        """
-        Adjustment along the longitudes for one specific latitude
-        used by :func:`cmethods.CMethods.adjust_3d`
-        as callback function for :class:`multiprocessing.Pool`.
-
-        **Not intended to be executed somewhere else.**
-
-        :params params: The method specific parameters
-        :type params: dict
-        :raises UnknownMethodError: If the specified method is not implemented
-        :return: The bias-corrected time series as 2-dimensional (longitudes x time)
-            numpy array
-        :rtype: np.ndarray
-        """
-        kwargs = params.get("kwargs", {})
-
-        result = params["simp"].copy(deep=True).load()
-        len_lon = len(params["obs"].lon)
-
-        func_adjustment = None
-        if params["method"] in cls.CUSTOM_METHODS:
-            func_adjustment = cls.get_function(params["method"])
-            kwargs["n_quantiles"] = params.get("n_quantiles", 100)
-            kwargs["kind"] = params.get("kind", "+")
-            for lon in range(len_lon):
-                result[lon] = func_adjustment(
-                    obs=params["obs"][lon],
-                    simh=params["simh"][lon],
-                    simp=params["simp"][lon],
-                    group=params.get("group", None),
-                    **kwargs,
-                )
-            return np.array(result)
-
-        raise UnknownMethodError(params["method"], cls.METHODS)
-
-    @classmethod
-    def grouped_correction(
-        cls: CMethods,
-        method: str,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
-        group: str,
-        kind: str = "+",
-        **kwargs: Any,
-    ) -> xr.core.dataarray.DataArray:
-        """Method to adjust 1 dimensional climate data while respecting adjustment groups.
-
-        ----- P A R A M E T E R S -----
-            method (str): adjustment method name
-            obs (xarray.core.dataarray.DataArray): observed / reference Data
-            simh (xarray.core.dataarray.DataArray): simulated historical Data
-            simp (xarray.core.dataarray.DataArray): future simulated Data
-            method (str): Scaling method (e.g.: 'linear_scaling')
-            group (str): [optional]  Group / Period (either: 'time.month', 'time.year' or 'time.dayofyear')
-
-        ----- R E T U R N -----
-
-            xarray.core.dataarray.DataArray: Adjusted data
-        """
-        if not isinstance(simp, xr.core.dataarray.DataArray):
-            raise TypeError("'simp' must be type xarray.core.dataarray.DataArray")
-
-        func_adjustment = cls.get_function(method)
-        result = simp.copy(deep=True).load()
-        groups = simh.groupby(group).groups
-
-        for month in groups.keys():
-            m_obs, m_simh, m_simp = [], [], []
-
-            for i in groups[month]:
-                m_obs.append(obs[i])
-                m_simh.append(simh[i])
-                m_simp.append(simp[i])
-
-            computed_result = func_adjustment(
-                obs=m_obs, simh=m_simh, simp=m_simp, kind=kind, group=None, **kwargs
-            )
-            for i, index in enumerate(groups[month]):
-                result[index] = computed_result[i]
-
-        return np.array(result)
 
     # ? -----========= L I N E A R - S C A L I N G =========------
     @classmethod
     def linear_scaling(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
-        group: str = "time.month",
+        cls,
+        obs,
+        simh,
+        simp,
         kind: str = "+",
         **kwargs: Any,
     ) -> np.ndarray:
@@ -420,7 +89,6 @@ class CMethods:
 
                 X^{*LS}_{sim,p}(i) = X_{sim,p}(i) + \mu_{m}(X_{obs,h}(i)) - \mu_{m}(X_{sim,h}(i))
 
-
         **Multiplicative**:
 
             The multiplicative Linear Scaling differs from the additive variant in such way, that the changes are not computed
@@ -430,7 +98,6 @@ class CMethods:
 
                 X^{*LS}_{sim,h}(i) = X_{sim,h}(i) \cdot \left[\frac{\mu_{m}(X_{obs,h}(i))}{\mu_{m}(X_{sim,h}(i))}\right]
 
-
         :param obs: The reference data set of the control period
             (in most cases the observational data)
         :type obs: xr.core.dataarray.DataArray
@@ -439,8 +106,6 @@ class CMethods:
         :param simp: The modeled data of the scenario period (this is the data set
             on which the bias correction takes action)
         :type simp: xr.core.dataarray.DataArray
-        :param group: The grouping defines the basis of the mean, defaults to ``time.month``
-        :type group: str | None
         :param kind: The kind of the correction, additive for non-stochastic and multiplicative
             for stochastic climate variables, defaults to ``+``
         :type kind: str, optional
@@ -469,24 +134,13 @@ class CMethods:
             ...     kind="+"
             ... )
         """
-        cls.check_types(obs=obs, simh=simh, simp=simp)
-
-        if group is not None:
-            return cls.grouped_correction(
-                method="linear_scaling",
-                obs=obs,
-                simh=simh,
-                simp=simp,
-                group=group,
-                kind=kind,
-                **kwargs,
-            )
-
-        if kind in cls.ADDITIVE:
+        if kind in ADDITIVE:
             return np.array(simp) + (np.nanmean(obs) - np.nanmean(simh))  # Eq. 1
-        if kind in cls.MULTIPLICATIVE:
-            adj_scaling_factor = cls.get_adjusted_scaling_factor(
-                cls.ensure_devidable(np.nanmean(obs), np.nanmean(simh)),
+        if kind in MULTIPLICATIVE:
+            adj_scaling_factor = get_adjusted_scaling_factor(
+                ensure_devidable(
+                    np.nanmean(obs), np.nanmean(simh), cls.MAX_SCALING_FACTOR
+                ),
                 kwargs.get("max_scaling_factor", cls.MAX_SCALING_FACTOR),
             )
             return np.array(simp) * adj_scaling_factor  # Eq. 2
@@ -497,11 +151,10 @@ class CMethods:
     # ? -----========= V A R I A N C E - S C A L I N G =========------
     @classmethod
     def variance_scaling(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
-        group: Optional[str] = "time.month",
+        cls,
+        obs,
+        simh,
+        simp,
         kind: str = "+",
         **kwargs: Any,
     ) -> np.ndarray:
@@ -529,7 +182,6 @@ class CMethods:
 
             X^{*LS}_{sim,p}(i) = X_{sim,p}(i) + \mu_{m}(X_{obs,h}(i)) - \mu_{m}(X_{sim,h}(i))
 
-
         **(2)** In the second step, the time-series are shifted to a zero mean. This enables the adjustment
         of the standard deviation in the following step.
 
@@ -539,20 +191,17 @@ class CMethods:
 
             X^{VS(1)}_{sim,p}(i) = X^{*LS}_{sim,p}(i) - \mu_{m}(X^{*LS}_{sim,p}(i))
 
-
         **(3)** Now the standard deviation (so variance too) can be scaled.
 
         .. math::
 
             X^{VS(2)}_{sim,p}(i) = X^{VS(1)}_{sim,p}(i) \cdot \left[\frac{\sigma_{m}(X_{obs,h}(i))}{\sigma_{m}(X^{VS(1)}_{sim,h}(i))}\right]
 
-
         **(4)** Finally the previously removed mean is shifted back
 
         .. math::
 
             X^{*VS}_{sim,p}(i) = X^{VS(2)}_{sim,p}(i) + \mu_{m}(X^{*LS}_{sim,p}(i))
-
 
         :param obs: The reference data set of the control period
             (in most cases the observational data)
@@ -562,8 +211,6 @@ class CMethods:
         :param simp: The modeled data of the scenario period (this is the data set
             on which the bias correction takes action)
         :type simp: xr.core.dataarray.DataArray
-        :param group: The grouping defines the basis of the mean, defaults to ``time.month``
-        :type group: str | None
         :param kind: The kind of the correction, additive for non-stochastic climate variables
             no other kind is available so far, defaults to ``+``
         :type kind: str, optional
@@ -591,28 +238,17 @@ class CMethods:
             ...     simp=simp[variable],
             ... )
         """
-        cls.check_types(obs=obs, simh=simh, simp=simp)
-
-        if group is not None:
-            return cls.grouped_correction(
-                method="variance_scaling",
-                obs=obs,
-                simh=simh,
-                simp=simp,
-                group=group,
-                kind=kind,
-                **kwargs,
-            )
-
-        if kind in cls.ADDITIVE:
-            LS_simh = cls.linear_scaling(obs, simh, simh, group=None)  # Eq. 1
-            LS_simp = cls.linear_scaling(obs, simh, simp, group=None)  # Eq. 2
+        if kind in ADDITIVE:
+            LS_simh = cls.linear_scaling(obs, simh, simh)  # Eq. 1
+            LS_simp = cls.linear_scaling(obs, simh, simp)  # Eq. 2
 
             VS_1_simh = LS_simh - np.nanmean(LS_simh)  # Eq. 3
             VS_1_simp = LS_simp - np.nanmean(LS_simp)  # Eq. 4
 
-            adj_scaling_factor = cls.get_adjusted_scaling_factor(
-                cls.ensure_devidable(np.std(np.array(obs)), np.std(VS_1_simh)),
+            adj_scaling_factor = get_adjusted_scaling_factor(
+                ensure_devidable(
+                    np.std(np.array(obs)), np.std(VS_1_simh), cls.MAX_SCALING_FACTOR
+                ),
                 kwargs.get("max_scaling_factor", cls.MAX_SCALING_FACTOR),
             )
 
@@ -626,11 +262,10 @@ class CMethods:
     # ? -----========= D E L T A - M E T H O D =========------
     @classmethod
     def delta_method(
-        cls: CMethods,
+        cls,
         obs: xr.core.dataarray.DataArray,
         simh: xr.core.dataarray.DataArray,
         simp: xr.core.dataarray.DataArray,
-        group: Optional[str] = "time.month",
         kind: str = "+",
         **kwargs: Any,
     ) -> np.ndarray:
@@ -665,7 +300,6 @@ class CMethods:
 
                 X^{*DM}_{sim,p}(i) = X_{obs,h}(i) + \mu_{m}(X_{sim,p}(i)) - \mu_{m}(X_{sim,h}(i))
 
-
         **Multiplicative**:
 
             The multiplicative variant behaves like the additive, but with the difference that the change is computed using the relative change
@@ -675,7 +309,6 @@ class CMethods:
 
                 X^{*DM}_{sim,p}(i) = X_{obs,h}(i) \cdot \left[\frac{ \mu_{m}(X_{sim,p}(i)) }{ \mu_{m}(X_{sim,h}(i))}\right]
 
-
         :param obs: The reference data set of the control period
             (in most cases the observational data)
         :type obs: xr.core.dataarray.DataArray
@@ -684,8 +317,6 @@ class CMethods:
         :param simp: The modeled data of the scenario period (this is the data set
             on which the bias correction takes action)
         :type simp: xr.core.dataarray.DataArray
-        :param group: The grouping defines the basis of the mean, defaults to ``time.month``
-        :type group: str | None
         :param kind: The kind of the correction, additive for non-stochastic and multiplicative
             for stochastic climate variables, defaults to ``+``
         :type kind: str, optional
@@ -713,24 +344,13 @@ class CMethods:
             ...     simp=simp[variable],
             ... )
         """
-        cls.check_types(obs=obs, simh=simh, simp=simp)
-
-        if group is not None:
-            return cls.grouped_correction(
-                method="delta_method",
-                obs=obs,
-                simh=simh,
-                simp=simp,
-                group=group,
-                kind=kind,
-                **kwargs,
-            )
-
-        if kind in cls.ADDITIVE:
+        if kind in ADDITIVE:
             return np.array(obs) + (np.nanmean(simp) - np.nanmean(simh))  # Eq. 1
-        if kind in cls.MULTIPLICATIVE:
-            adj_scaling_factor = cls.get_adjusted_scaling_factor(
-                cls.ensure_devidable(np.nanmean(simp), np.nanmean(simh)),
+        if kind in MULTIPLICATIVE:
+            adj_scaling_factor = get_adjusted_scaling_factor(
+                ensure_devidable(
+                    np.nanmean(simp), np.nanmean(simh), cls.MAX_SCALING_FACTOR
+                ),
                 kwargs.get("max_scaling_factor", cls.MAX_SCALING_FACTOR),
             )
             return np.array(obs) * adj_scaling_factor  # Eq. 2
@@ -741,10 +361,10 @@ class CMethods:
     # ? -----========= Q U A N T I L E - M A P P I N G =========------
     @classmethod
     def quantile_mapping(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
+        cls,
+        obs,
+        simh,
+        simp,
         n_quantiles: int,
         kind: str = "+",
         **kwargs: Any,
@@ -849,7 +469,7 @@ class CMethods:
             ...     n_quantiles=250
             ... )
         """
-        cls.check_types(obs=obs, simh=simh, simp=simp)
+        check_types(obs=obs, simh=simh, simp=simp)
         if not isinstance(n_quantiles, int):
             raise TypeError("'n_quantiles' must be type int")
 
@@ -858,20 +478,20 @@ class CMethods:
         global_max = max(np.nanmax(obs), np.nanmax(simh))
         global_min = min(np.nanmin(obs), np.nanmin(simh))
 
-        if cls.nan_or_equal(value1=global_max, value2=global_min):
+        if nan_or_equal(value1=global_max, value2=global_min):
             return simp
 
         wide = abs(global_max - global_min) / n_quantiles
         xbins = np.arange(global_min, global_max + wide, wide)
 
-        cdf_obs = cls.get_cdf(obs, xbins)
-        cdf_simh = cls.get_cdf(simh, xbins)
+        cdf_obs = get_cdf(obs, xbins)
+        cdf_simh = get_cdf(simh, xbins)
 
-        if kind in cls.ADDITIVE:
+        if kind in ADDITIVE:
             epsilon = np.interp(simp, xbins, cdf_simh)  # Eq. 1
-            return cls.get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1
+            return get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1
 
-        if kind in cls.MULTIPLICATIVE:
+        if kind in MULTIPLICATIVE:
             epsilon = np.interp(  # Eq. 2
                 simp,
                 xbins,
@@ -879,169 +499,173 @@ class CMethods:
                 left=kwargs.get("val_min", 0.0),
                 right=kwargs.get("val_max", None),
             )
-            return cls.get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 2
+            return get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 2
 
         raise NotImplementedError(
             f"{kind} for quantile_mapping is not available. Use '+' or '*' instead."
         )
 
-    @classmethod
-    def detrended_quantile_mapping(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
-        n_quantiles: int,
-        kind: str = "+",
-        **kwargs: Any,
-    ) -> np.ndarray:
-        r"""
-        The Detrended Quantile Mapping bias correction technique can be used to minimize distributional
-        biases between modeled and observed time-series climate data like the regular Quantile Mapping.
-        Detrending means, that the values of :math:`X_{sim,p}` are shifted to the value range of
-        :math:`X_{sim,h}` before the regular Quantile Mapping is applied.
-        After the Quantile Mapping was applied, the mean is shifted back. Since it does not make sens
-        to take the whole mean to rescale the data, the month-dependent long-term mean is used.
+    # @classmethod
+    # def detrended_quantile_mapping(
+    #     cls,
+    #     obs,
+    #     simh,
+    #     simp,
+    #     n_quantiles: int,
+    #     kind: str = "+",
+    #     **kwargs: Any,
+    # ) -> np.ndarray:
+    #     r"""
+    #     The Detrended Quantile Mapping bias correction technique can be used to minimize distributional
+    #     biases between modeled and observed time-series climate data like the regular Quantile Mapping.
+    #     Detrending means, that the values of :math:`X_{sim,p}` are shifted to the value range of
+    #     :math:`X_{sim,h}` before the regular Quantile Mapping is applied.
+    #     After the Quantile Mapping was applied, the mean is shifted back. Since it does not make sens
+    #     to take the whole mean to rescale the data, the month-dependent long-term mean is used.
 
-        This method must be applied on a 1-dimensional data set i.e., there is only one
-        time-series passed for each of ``obs``, ``simh``, and ``simp``. Also this method requires
-        that the time series are groupable by ``time.month``.
+    #     This method must be applied on a 1-dimensional data set i.e., there is only one
+    #     time-series passed for each of ``obs``, ``simh``, and ``simp``. Also this method requires
+    #     that the time series are groupable by ``time.month``.
 
-        The Detrended Quantile Mapping technique implemented here is based on the equations of
-        Alex J. Cannon and Stephen R. Sobie and Trevor Q. Murdock (2015) *"Bias Correction of GCM
-        Precipitation by Quantile Mapping: How Well Do Methods Preserve Changes in Quantiles
-        and Extremes?"* (https://doi.org/10.1175/JCLI-D-14-00754.1).
+    #     The Detrended Quantile Mapping technique implemented here is based on the equations of
+    #     Alex J. Cannon and Stephen R. Sobie and Trevor Q. Murdock (2015) *"Bias Correction of GCM
+    #     Precipitation by Quantile Mapping: How Well Do Methods Preserve Changes in Quantiles
+    #     and Extremes?"* (https://doi.org/10.1175/JCLI-D-14-00754.1).
 
-        In the following the equations of Alex J. Cannon (2015) are shown (without detrending; see QM
-        for explanations):
+    #     In the following the equations of Alex J. Cannon (2015) are shown (without detrending; see QM
+    #     for explanations):
 
-        **Additive**:
+    #     **Additive**:
 
-            .. math::
+    #         .. math::
 
-                X^{*QM}_{sim,p}(i) = F^{-1}_{obs,h} \left\{F_{sim,h}\left[X_{sim,p}(i)\right]\right\}
+    #             X^{*QM}_{sim,p}(i) = F^{-1}_{obs,h} \left\{F_{sim,h}\left[X_{sim,p}(i)\right]\right\}
 
+    #     **Multiplicative**:
 
-        **Multiplicative**:
+    #         .. math::
 
-            .. math::
+    #             X^{*QM}_{sim,p}(i) = F^{-1}_{obs,h}\Biggl\{F_{sim,h}\left[\frac{\mu{X_{sim,h}} \cdot \mu{X_{sim,p}(i)}}{\mu{X_{sim,p}(i)}}\right]\Biggr\}\frac{\mu{X_{sim,p}(i)}}{\mu{X_{sim,h}}}
 
-                X^{*QM}_{sim,p}(i) = F^{-1}_{obs,h}\Biggl\{F_{sim,h}\left[\frac{\mu{X_{sim,h}} \cdot \mu{X_{sim,p}(i)}}{\mu{X_{sim,p}(i)}}\right]\Biggr\}\frac{\mu{X_{sim,p}(i)}}{\mu{X_{sim,h}}}
+    #     :param obs: The reference data set of the control period
+    #         (in most cases the observational data)
+    #     :type obs: xr.core.dataarray.DataArray
+    #     :param simh: The modeled data of the control period
+    #     :type simh: xr.core.dataarray.DataArray
+    #     :param simp: The modeled data of the scenario period (this is the data set
+    #         on which the bias correction takes action)
+    #     :type simp: xr.core.dataarray.DataArray
+    #     :param n_quantiles: Number of quantiles to respect/use
+    #     :type n_quantiles: int
+    #     :param kind: The kind of the correction, additive for non-stochastic and multiplicative
+    #         for stochastic climate variables, defaults to ``+``
+    #     :type kind: str, optional
+    #     :param val_min: Lower boundary for interpolation (only if ``kind="*"``, default: ``0.0``)
+    #     :type val_min: float, optional
+    #     :param val_max: Upper boundary for interpolation (only if ``kind="*"``, default: ``None``)
+    #     :type val_max: float, optional
+    #     :raises NotImplementedError: If the kind is not in (``+``, ``*``, ``add``, ``mult``)
+    #     :return: The bias-corrected time series
+    #     :rtype: np.ndarray
 
+    #     .. code-block:: python
+    #         :linenos:
+    #         :caption: Example: Quantile Mapping
 
-        :param obs: The reference data set of the control period
-            (in most cases the observational data)
-        :type obs: xr.core.dataarray.DataArray
-        :param simh: The modeled data of the control period
-        :type simh: xr.core.dataarray.DataArray
-        :param simp: The modeled data of the scenario period (this is the data set
-            on which the bias correction takes action)
-        :type simp: xr.core.dataarray.DataArray
-        :param n_quantiles: Number of quantiles to respect/use
-        :type n_quantiles: int
-        :param kind: The kind of the correction, additive for non-stochastic and multiplicative
-            for stochastic climate variables, defaults to ``+``
-        :type kind: str, optional
-        :param val_min: Lower boundary for interpolation (only if ``kind="*"``, default: ``0.0``)
-        :type val_min: float, optional
-        :param val_max: Upper boundary for interpolation (only if ``kind="*"``, default: ``None``)
-        :type val_max: float, optional
-        :raises NotImplementedError: If the kind is not in (``+``, ``*``, ``add``, ``mult``)
-        :return: The bias-corrected time series
-        :rtype: np.ndarray
+    #         >>> import xarray as xr
+    #         >>> from cmethods import CMethods as cm
 
-        .. code-block:: python
-            :linenos:
-            :caption: Example: Quantile Mapping
+    #         >>> # Note: The data sets must contain the dimension "time"
+    #         >>> #       for the respective variable.
+    #         >>> obsh = xr.open_dataset("path/to/reference_data-control_period.nc")
+    #         >>> simh = xr.open_dataset("path/to/modeled_data-control_period.nc")
+    #         >>> simp = xr.open_dataset("path/to/the_dataset_to_adjust-scenario_period.nc")
+    #         >>> variable = "tas" # temperatures
 
-            >>> import xarray as xr
-            >>> from cmethods import CMethods as cm
+    #         >>> qm_adjusted = cm.detrended_quantile_mapping(
+    #         ...     obs=obs[variable],
+    #         ...     simh=simh[variable],
+    #         ...     simp=simp[variable],
+    #         ...     n_quantiles=250
+    #         ... )
+    #     """
+    #     if kind not in MULTIPLICATIVE and kind not in ADDITIVE:
+    #         raise NotImplementedError(
+    #             f"{kind} for detrended_quantile_mapping is not available. Use '+' or '*' instead."
+    #         )
 
-            >>> # Note: The data sets must contain the dimension "time"
-            >>> #       for the respective variable.
-            >>> obsh = xr.open_dataset("path/to/reference_data-control_period.nc")
-            >>> simh = xr.open_dataset("path/to/modeled_data-control_period.nc")
-            >>> simp = xr.open_dataset("path/to/the_dataset_to_adjust-scenario_period.nc")
-            >>> variable = "tas" # temperatures
+    #     check_types(obs=obs, simh=simh, simp=simp)
+    #     if not isinstance(n_quantiles, int):
+    #         raise TypeError("'n_quantiles' must be type int")
+    #     if not isinstance(simp, xr.core.dataarray.DataArray):
+    #         raise TypeError("'simp' must be type xarray.core.dataarray.DataArray")
 
-            >>> qm_adjusted = cm.detrended_quantile_mapping(
-            ...     obs=obs[variable],
-            ...     simh=simh[variable],
-            ...     simp=simp[variable],
-            ...     n_quantiles=250
-            ... )
-        """
-        if kind not in cls.MULTIPLICATIVE and kind not in cls.ADDITIVE:
-            raise NotImplementedError(
-                f"{kind} for detrended_quantile_mapping is not available. Use '+' or '*' instead."
-            )
+    #     obs, simh = np.array(obs), np.array(simh)
 
-        cls.check_types(obs=obs, simh=simh, simp=simp)
-        if not isinstance(n_quantiles, int):
-            raise TypeError("'n_quantiles' must be type int")
-        if not isinstance(simp, xr.core.dataarray.DataArray):
-            raise TypeError("'simp' must be type xarray.core.dataarray.DataArray")
+    #     global_max = max(np.nanmax(obs), np.nanmax(simh))
+    #     global_min = min(np.nanmin(obs), np.nanmin(simh))
 
-        obs, simh = np.array(obs), np.array(simh)
+    #     if nan_or_equal(value1=global_max, value2=global_min):
+    #         return np.array(simp.values)
 
-        global_max = max(np.nanmax(obs), np.nanmax(simh))
-        global_min = min(np.nanmin(obs), np.nanmin(simh))
+    #     wide = abs(global_max - global_min) / n_quantiles
+    #     xbins = np.arange(global_min, global_max + wide, wide)
 
-        if cls.nan_or_equal(value1=global_max, value2=global_min):
-            return np.array(simp.values)
+    #     cdf_obs = get_cdf(obs, xbins)
+    #     cdf_simh = get_cdf(simh, xbins)
 
-        wide = abs(global_max - global_min) / n_quantiles
-        xbins = np.arange(global_min, global_max + wide, wide)
+    #     # detrended => shift mean of $X_{sim,p}$ to range of $X_{sim,h}$ to adjust extremes
+    #     res = np.zeros(len(simp.values))
+    #     for indices in simp.groupby("time.month").groups.values():
+    #         # detrended by long-term month
+    #         m_simh, m_simp = [], []
+    #         for index in indices:
+    #             m_simh.append(simh[index])
+    #             m_simp.append(simp[index])
 
-        cdf_obs = cls.get_cdf(obs, xbins)
-        cdf_simh = cls.get_cdf(simh, xbins)
+    #         m_simh = np.array(m_simh)
+    #         m_simp = np.array(m_simp)
+    #         m_simh_mean = np.nanmean(m_simh)
+    #         m_simp_mean = np.nanmean(m_simp)
 
-        # detrended => shift mean of $X_{sim,p}$ to range of $X_{sim,h}$ to adjust extremes
-        res = np.zeros(len(simp.values))
-        for indices in simp.groupby("time.month").groups.values():
-            # detrended by long-term month
-            m_simh, m_simp = [], []
-            for index in indices:
-                m_simh.append(simh[index])
-                m_simp.append(simp[index])
+    #         if kind in ADDITIVE:
+    #             epsilon = np.interp(
+    #                 m_simp - m_simp_mean + m_simh_mean, xbins, cdf_simh
+    #             )  # Eq. 1
+    #             X = (
+    #                 get_inverse_of_cdf(cdf_obs, epsilon, xbins)
+    #                 + m_simp_mean
+    #                 - m_simh_mean
+    #             )  # Eq. 1
 
-            m_simh = np.array(m_simh)
-            m_simp = np.array(m_simp)
-            m_simh_mean = np.nanmean(m_simh)
-            m_simp_mean = np.nanmean(m_simp)
-
-            if kind in cls.ADDITIVE:
-                epsilon = np.interp(
-                    m_simp - m_simp_mean + m_simh_mean, xbins, cdf_simh
-                )  # Eq. 1
-                X = (
-                    cls.get_inverse_of_cdf(cdf_obs, epsilon, xbins)
-                    + m_simp_mean
-                    - m_simh_mean
-                )  # Eq. 1
-
-            else:  # kind in cls.MULTIPLICATIVE:
-                epsilon = np.interp(  # Eq. 2
-                    cls.ensure_devidable((m_simh_mean * m_simp), m_simp_mean),
-                    xbins,
-                    cdf_simh,
-                    left=kwargs.get("val_min", 0.0),
-                    right=kwargs.get("val_max", None),
-                )
-                X = np.interp(epsilon, cdf_obs, xbins) * cls.ensure_devidable(
-                    m_simp_mean, m_simh_mean
-                )  # Eq. 2
-            for i, index in enumerate(indices):
-                res[index] = X[i]
-        return res
+    #         else:  # kind in MULTIPLICATIVE:
+    #             epsilon = np.interp(  # Eq. 2
+    #                 ensure_devidable(
+    #                     (m_simh_mean * m_simp),
+    #                     m_simp_mean,
+    #                     max_scaling_factor=cls.MAX_SCALING_FACTOR,
+    #                 ),
+    #                 xbins,
+    #                 cdf_simh,
+    #                 left=kwargs.get("val_min", 0.0),
+    #                 right=kwargs.get("val_max", None),
+    #             )
+    #             X = np.interp(epsilon, cdf_obs, xbins) * ensure_devidable(
+    #                 m_simp_mean,
+    #                 m_simh_mean,
+    #                 max_scaling_factor=cls.MAX_SCALING_FACTOR,
+    #             )  # Eq. 2
+    #         for i, index in enumerate(indices):
+    #             res[index] = X[i]
+    #     return res
 
     # ? -----========= E M P I R I C A L - Q U A N T I L E - M A P P I N G =========------
     @classmethod
     def empirical_quantile_mapping(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
+        cls,
+        obs,
+        simh,
+        simp,
         n_quantiles: int = 100,
         extrapolate: Optional[str] = None,
         **kwargs: Any,
@@ -1073,10 +697,10 @@ class CMethods:
     # ? -----========= Q U A N T I L E - D E L T A - M A P P I N G =========------
     @classmethod
     def quantile_delta_mapping(
-        cls: CMethods,
-        obs: xr.core.dataarray.DataArray,
-        simh: xr.core.dataarray.DataArray,
-        simp: xr.core.dataarray.DataArray,
+        cls,
+        obs,
+        simh,
+        simp,
         n_quantiles: int,
         kind: str = "+",
         **kwargs: Any,
@@ -1108,7 +732,6 @@ class CMethods:
 
                 \varepsilon(i) = F_{sim,p}\left[X_{sim,p}(i)\right], \hspace{1em} \varepsilon(i)\in\{0,1\}
 
-
             **(1.2)** The bias corrected value at time step :math:`i` is now determined by inserting the
             quantile value into the inverse cummulative distribution function of the reference data of the control
             period. This results in a bias corrected value for time step :math:`i` but still without taking the
@@ -1118,7 +741,6 @@ class CMethods:
 
                 X^{QDM(1)}_{sim,p}(i) = F^{-1}_{obs,h}\left[\varepsilon(i)\right]
 
-
             **(1.3)** The :math:`\Delta(i)` represents the absolute change in quantiles between the modeled value
             in the control and scenario period.
 
@@ -1127,13 +749,11 @@ class CMethods:
                  \Delta(i) & = F^{-1}_{sim,p}\left[\varepsilon(i)\right] - F^{-1}_{sim,h}\left[\varepsilon(i)\right] \\[1pt]
                            & = X_{sim,p}(i) - F^{-1}_{sim,h}\left\{F^{}_{sim,p}\left[X_{sim,p}(i)\right]\right\}
 
-
             **(1.4)** Finally the previously calculated change can be added to the bias-corrected value.
 
             .. math::
 
                 X^{*QDM}_{sim,p}(i) = X^{QDM(1)}_{sim,p}(i) + \Delta(i)
-
 
         **Multiplicative**:
 
@@ -1148,14 +768,12 @@ class CMethods:
                     \Delta(i) & = \frac{ F^{-1}_{sim,p}\left[\varepsilon(i)\right] }{ F^{-1}_{sim,h}\left[\varepsilon(i)\right] } \\[1pt]
                               & = \frac{ X_{sim,p}(i) }{ F^{-1}_{sim,h}\left\{F_{sim,p}\left[X_{sim,p}(i)\right]\right\} }
 
-
             **(2.4)** The relative change between the modeled data of the control and scenario period is than
             multiplicated with the bias-corrected value (see **1.2**).
 
                 .. math::
 
                     X^{*QDM}_{sim,p}(i) = X^{QDM(1)}_{sim,p}(i) \cdot \Delta(i)
-
 
         :param obs: The reference data set of the control period
             (in most cases the observational data)
@@ -1195,12 +813,12 @@ class CMethods:
             ...     n_quantiles=250
             ... )
         """
-        cls.check_types(obs=obs, simh=simh, simp=simp)
+        check_types(obs=obs, simh=simh, simp=simp)
 
         if not isinstance(n_quantiles, int):
             raise TypeError("'n_quantiles' must be type int")
 
-        if kind in cls.ADDITIVE:
+        if kind in ADDITIVE:
             obs, simh, simp = (
                 np.array(obs),
                 np.array(simh),
@@ -1209,41 +827,43 @@ class CMethods:
             global_max = kwargs.get("global_max", max(np.nanmax(obs), np.nanmax(simh)))
             global_min = kwargs.get("global_min", min(np.nanmin(obs), np.nanmin(simh)))
 
-            if cls.nan_or_equal(value1=global_max, value2=global_min):
+            if nan_or_equal(value1=global_max, value2=global_min):
                 return simp
 
             wide = abs(global_max - global_min) / n_quantiles
             xbins = np.arange(global_min, global_max + wide, wide)
 
-            cdf_obs = cls.get_cdf(obs, xbins)
-            cdf_simh = cls.get_cdf(simh, xbins)
-            cdf_simp = cls.get_cdf(simp, xbins)
+            cdf_obs = get_cdf(obs, xbins)
+            cdf_simh = get_cdf(simh, xbins)
+            cdf_simp = get_cdf(simp, xbins)
 
             # calculate exact cdf values of $F_{sim,p}[T_{sim,p}(t)]$
             epsilon = np.interp(simp, xbins, cdf_simp)  # Eq. 1.1
-            QDM1 = cls.get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1.2
-            delta = simp - cls.get_inverse_of_cdf(cdf_simh, epsilon, xbins)  # Eq. 1.3
+            QDM1 = get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1.2
+            delta = simp - get_inverse_of_cdf(cdf_simh, epsilon, xbins)  # Eq. 1.3
             return QDM1 + delta  # Eq. 1.4
 
-        if kind in cls.MULTIPLICATIVE:
+        if kind in MULTIPLICATIVE:
             obs, simh, simp = np.array(obs), np.array(simh), np.array(simp)
             global_max = kwargs.get("global_max", max(np.nanmax(obs), np.nanmax(simh)))
             global_min = kwargs.get("global_min", 0.0)
-            if cls.nan_or_equal(value1=global_max, value2=global_min):
+            if nan_or_equal(value1=global_max, value2=global_min):
                 return simp
 
             wide = global_max / n_quantiles
             xbins = np.arange(global_min, global_max + wide, wide)
 
-            cdf_obs = cls.get_cdf(obs, xbins)
-            cdf_simh = cls.get_cdf(simh, xbins)
-            cdf_simp = cls.get_cdf(simp, xbins)
+            cdf_obs = get_cdf(obs, xbins)
+            cdf_simh = get_cdf(simh, xbins)
+            cdf_simp = get_cdf(simp, xbins)
 
             epsilon = np.interp(simp, xbins, cdf_simp)  # Eq. 1.1
-            QDM1 = cls.get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1.2
+            QDM1 = get_inverse_of_cdf(cdf_obs, epsilon, xbins)  # Eq. 1.2
 
-            delta = cls.ensure_devidable(  # Eq. 2.3
-                simp, cls.get_inverse_of_cdf(cdf_simh, epsilon, xbins)
+            delta = ensure_devidable(  # Eq. 2.3
+                simp,
+                get_inverse_of_cdf(cdf_simh, epsilon, xbins),
+                max_scaling_factor=cls.MAX_SCALING_FACTOR,
             )
             return QDM1 * delta  # Eq. 2.4
         raise NotImplementedError(
@@ -1251,176 +871,103 @@ class CMethods:
         )
 
     @classmethod
-    def check_types(
-        cls: CMethods,
-        obs: Union[list, np.ndarray, np.generic, xr.core.dataarray.DataArray],
-        simh: Union[list, np.ndarray, np.generic, xr.core.dataarray.DataArray],
-        simp: Union[list, np.ndarray, np.generic, xr.core.dataarray.DataArray],
-    ) -> None:
-        """
-        Checks if the parameters are in the correct type. **only used internally**
-        """
-        phrase: str = "must be type list, np.ndarray or xarray.core.dataarray.DataArray"
-        valid_types: tuple = (list, np.ndarray, np.generic, xr.core.dataarray.DataArray)
+    def adjust(cls, method, obs, simh, simp, **kwargs):
+        if kwargs.get("group", None) is not None:
+            group = kwargs["group"]
+            del kwargs["group"]
 
-        if not isinstance(obs, valid_types):
-            raise TypeError(f"'obs' {phrase}")
-        if not isinstance(simh, valid_types):
-            raise TypeError(f"'simh' {phrase}")
-        if not isinstance(simp, valid_types):
-            raise TypeError(f"'simp' {phrase}")
+            obs_g = list(obs.groupby(group))
+            simh_g = list(simh.groupby(group))
+            simp_g = list(simp.groupby(group))
 
-    @staticmethod
-    def nan_or_equal(value1: float, value2: float) -> bool:
-        """
-        Returns True if the values are equal or at least one is NaN
+            result = None
+            for index in range(len(list(obs_g))):
+                obs_gds = obs_g[index][1]
+                simh_gds = simh_g[index][1]
+                simp_gds = simp_g[index][1]
 
-        :param value1: First value to check
-        :type value1: float
-        :param value2: Second value to check
-        :type value2: float
-        :return: If any value is NaN or values are equal
-        :rtype: bool
-        """
-        return np.isnan(value1) or np.isnan(value2) or value1 == value2
+                monthly_result = cls.apply_ufunc(
+                    method, obs_gds, simh_gds, simp_gds, **kwargs
+                )
+                if result is None:
+                    result = monthly_result
+                else:
+                    result = xr.concat([result, monthly_result], dim="time")
+            return result
+        return cls.apply_ufunc(method, obs, simh, simp, **kwargs)
 
     @classmethod
-    def ensure_devidable(
-        cls: CMethods,
-        numerator: Union[float, np.ndarray],
-        denominator: Union[float, np.ndarray],
-    ) -> np.ndarray:
+    def get_function(cls, method):
         """
-        Ensures that the arrays can be devided. The numerator will be multiplied by
-        the maximum scaling factor of the CMethods class if division by zero.
+        Returns the bias correction function corresponding to the ``method`` name.
 
-        :param numerator: Numerator to use
-        :type numerator: np.ndarray
-        :param denominator: Denominator that can be zero
-        :type denominator: np.ndarray
-        :return: Zero-ensured devision
-        :rtype: np.ndarray | float
+        :param method: The method name to get the function for
+        :type method: str
+        :raises UnknownMethodError: If the function is not implemented
+        :return: The function of the corresponding method
+        :rtype: function
         """
-        with np.errstate(divide="ignore", invalid="ignore"):
-            result = numerator / denominator
+        if method == "linear_scaling":
+            return cls.linear_scaling
+        if method == "variance_scaling":
+            return cls.variance_scaling
+        if method == "delta_method":
+            return cls.delta_method
+        if method == "quantile_mapping":
+            return cls.quantile_mapping
+        if method == "detrended_quantile_mapping":
+            return cls.detrended_quantile_mapping
+        if method == "empirical_quantile_mapping":
+            return cls.empirical_quantile_mapping
+        if method == "quantile_delta_mapping":
+            return cls.quantile_delta_mapping
+        raise UnknownMethodError(method, METHODS)
 
-        if isinstance(numerator, np.ndarray):
-            mask_inf = np.isinf(result)
-            result[mask_inf] = numerator[mask_inf] * cls.MAX_SCALING_FACTOR
+    @classmethod
+    def apply_ufunc(cls, method, obs, simh, simp, **kwargs):
+        result = xr.apply_ufunc(
+            cls.get_function(method),
+            obs,
+            simh,
+            # Need to spoof a fake time axis since 'time' coord on full dataset is different
+            # than 'time' coord on training dataset.
+            simp.rename({"time": "t2"}),
+            dask="parallelized",
+            vectorize=True,
+            # This will vectorize over the time dimension, so will submit each grid cell
+            # independently
+            input_core_dims=[["time"], ["time"], ["t2"]],
+            # Need to denote that the final output dataset will be labeled with the
+            # spoofed time coordinate
+            output_core_dims=[["t2"]],
+            kwargs=dict(kwargs),
+        )
 
-            mask_nan = np.isnan(result)
-            result[mask_nan] = 0
-        elif np.isinf(result):
-            result = numerator * cls.MAX_SCALING_FACTOR
-        elif np.isnan(result):
-            result = 0.0
+        # Rename to proper coordinate name.
+        result = result.rename({"t2": "time"})
 
-        return result
+        # ufunc will put the core dimension to the end (time), so want to preserve original
+        # order where time is commonly first.
+        return result.transpose(*obs.dims)
 
-    @staticmethod
-    def get_pdf(
-        x: Union[list, np.ndarray], xbins: Union[list, np.ndarray]
-    ) -> np.ndarray:
-        r"""
-        Compuites and returns the the probability density function :math:`P(x)`
-        of ``x`` based on ``xbins``.
+    @classmethod
+    def get_available_methods(cls) -> List[str]:
+        """
+        Function to return the available adjustment methods of the CMethods class.
 
-        :param x: The vector to get :math:`P(x)` from
-        :type x: Union[list, np.ndarray]
-        :param xbins: The boundaries/bins of :math:`P(x)`
-        :type xbins: Union[list, np.ndarray]
-        :return: The probability densitiy function of ``x``
-        :rtype: np.ndarray
+        :return: List of available bias correction methods
+        :rtype: List[str]
 
         .. code-block:: python
             :linenos:
-            :caption: Compute the probability density function :math:`P(x)`
+            :caption: Example: Get available methods
 
             >>> from cmethods import CMethods as cm
 
-            >>> x = [1, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10]
-            >>> xbins = [0, 3, 6, 10]
-            >>> print(cm.get_pdf(x=x, xbins=xbins))
-            [2, 5, 5]
+            >>> cm.get_available_methods()
+            [
+                "linear_scaling", "variance_scaling", "delta_method",
+                "quantile_mapping", "quantile_delta_mapping"
+            ]
         """
-        pdf, _ = np.histogram(x, xbins)
-        return pdf
-
-    @staticmethod
-    def get_cdf(
-        x: Union[list, np.ndarray], xbins: Union[list, np.ndarray]
-    ) -> np.ndarray:
-        r"""
-        Computes and returns returns the cumulative distribution function :math:`F(x)`
-        of ``x`` based on ``xbins``.
-
-        :param x: Vector to get :math:`F(x)` from
-        :type x: Union[list, np.ndarray]
-        :param xbins: The boundaries/bins of :math:`F(x)`
-        :type xbins: Union[list, np.ndarray]
-        :return: The cumulative distribution function of ``x``
-        :rtype: np.ndarray
-
-
-        .. code-block:: python
-            :linenos:
-            :caption: Compute the cmmulative distribution function :math:`F(x)`
-
-            >>> from cmethods import CMethods as cm
-
-            >>> x = [1, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10]
-            >>> xbins = [0, 3, 6, 10]
-            >>> print(cm.get_cdf(x=x, xbins=xbins))
-            [0, 2, 7, 12]
-        """
-        pdf, _ = np.histogram(x, xbins)
-        return np.insert(np.cumsum(pdf), 0, 0.0)
-
-    @staticmethod
-    def get_inverse_of_cdf(
-        base_cdf: Union[list, np.ndarray],
-        insert_cdf: Union[list, np.ndarray],
-        xbins: Union[list, np.ndarray],
-    ) -> np.ndarray:
-        r"""
-        Returns the inverse cumulative distribution function as:
-        :math:`F^{-1}_{x}\left[y\right]` where :math:`x` represents ``base_cdf`` and
-        ``insert_cdf`` is represented by :math:`y`.
-
-        :param base_cdf: The basis
-        :type base_cdf: Union[list, np.ndarray]
-        :param insert_cdf: The CDF that gets inserted
-        :type insert_cdf: Union[list, np.ndarray]
-        :param xbins: Probability boundaries
-        :type xbins: Union[list, np.ndarray]
-        :return: The inverse CDF
-        :rtype: np.ndarray
-        """
-        return np.interp(insert_cdf, base_cdf, xbins)
-
-    @staticmethod
-    def get_adjusted_scaling_factor(
-        factor: Union[int, float], max_scaling_factor: Union[int, float]
-    ) -> float:
-        r"""
-        Returns:
-            - :math:`x` if :math:`-|y| \le x \le |y|`,
-            - :math:`|y|` if :math:`x > |y|`, or
-            - :math:`-|y|` if :math:`x < -|y|`
-
-            where:
-                - :math:`x` is ``factor``
-                - :math:`y` is ``max_scaling_factor``
-
-        :param factor: The value to check for
-        :type factor: Union[int, float]
-        :param max_scaling_factor: The maximum/minimum allowed value
-        :type max_scaling_factor: Union[int, float]
-        :return: The correct value
-        :rtype: float
-        """
-        if factor > 0 and factor > abs(max_scaling_factor):
-            return abs(max_scaling_factor)
-        if factor < 0 and factor < -abs(max_scaling_factor):
-            return -abs(max_scaling_factor)
-        return factor
+        return METHODS
